Generate types from Contentful content type
This uses contentful-typescript-codegen to do so.
The file is then manually pruned to remove excess types and fix the
weird I<interfaceName> naming.

Also, TODO: update the fork of contentful-typescript-codegen to be
more user friendly by choosing naming type, not including unused
imports like @contentful/rich-text-types and allowing the user to
select which content type to export the types of using a CLI flag.